% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{anovaP}
\alias{anovaP}
\title{ANOVA test P value}
\usage{
anovaP(expr, label, base = NULL, more = TRUE)
}
\arguments{
\item{expr}{a vector of continuous values}

\item{label}{a categorical vector or factor}

\item{base}{base for calculating Fold Change}

\item{more}{whether to supply just P value or add Tukey HSD info}
}
\value{
a vector
}
\description{
This function calculates F test P value and Tukey HSD test. The adjusted P values, difference of means as well as FC are also calculated
}

