% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{augmentMatrixColumnByID_update}
\alias{augmentMatrixColumnByID_update}
\title{update augMat after it is subsetted}
\usage{
augmentMatrixColumnByID_update(mat, ID, pNAthr = 0.8)
}
\arguments{
\item{ID}{IDs targeted for; should be simialr to colnames(mat)}

\item{computeNA}{whether to recompute columns with all NA}
}
\value{
augMat
}
\description{
update augMat after it is subsetted
}

