% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{augmentMatrixRowByID}
\alias{augmentMatrixRowByID}
\title{Augment a data frame by ID in the direction of rows}
\usage{
augmentMatrixRowByID(mat, ID, computeNA = FALSE, pNAthr = 0.8)
}
\arguments{
\item{mat}{a matrix or data frame to be augmented}

\item{ID}{IDs targeted for; should be simialr to rownames(mat)}

\item{computeNA}{whether to recompute columns with all NA}

\item{pNAthr}{percentage of NA to be called as NA columns}
}
\value{
an augMat object (also inherits matrix)
}
\description{
similar to fillVecByID, we might want to augment a matrix by column names specified by ID
We cannot directly using augmentMatrixColumnByID with transposed input since data frame has side effects converting all numeric to string
}

