% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{autocolpal4DF}
\alias{autocolpal4DF}
\title{get default colorpal (color vector) for a data frame (colwise)}
\usage{
autocolpal4DF(dat, alpha = NULL, economicBrewer = TRUE, pal = NULL)
}
\arguments{
\item{dat}{a data frame}

\item{pal}{optional user specified palette to update the final colpal}
}
\value{
a (named) list of color vectors that can be used to build colormap and colormat
}
\description{
this is auto color pallete; for better vis, the user need to specify it explicitely by modifying the returned list or start from scratch
}

