% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencpu.R
\name{buildapp}
\alias{buildapp}
\title{build an openCPU app with roxygenize+build+install+opencpu$browse}
\usage{
buildapp(package, reopen = TRUE)
}
\arguments{
\item{package}{package name, a string}

\item{reopen}{whether to open the browser}
}
\description{
this function wrappers roxygenize(), build(), install(), and browse()
}
\details{
Importance: need to navigate to the mater folder first; cannot work on full path!
}

