% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{categorizePvec}
\alias{categorizePvec}
\title{convert P value into categories}
\usage{
categorizePvec(Pvec)
}
\arguments{
\item{Pvec}{a vector of P value}
}
\value{
result by cut function
}
\description{
convert P value into categories
}

