% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{colorpalette2colvec}
\alias{colorpalette2colvec}
\title{extracts color given a specified pallete (string) possibly with alpha blending and number of colors desired}
\usage{
colorpalette2colvec(pal = "brewer.comb", nrcol = NULL, alpha = NULL,
  seed = NULL)
}
\arguments{
\item{pal}{string specifying pallete (predefined)}

\item{nrcol}{number of colors desired}

\item{alpha}{alpha blending}
}
\value{
a vector of colors
}
\description{
this function is originally from LSD but adapted a little (adding more colors). This function will return the color if nrcol
is not specified; otherwise, color interperlation through grDevices::colorRampPalette will be used.
}

