% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{combine_cormat}
\alias{combine_cormat}
\title{when pearson and spearman values both are desired, we can fit them separately and then combine}
\usage{
combine_cormat(fit1, fit2, tag = c("spearman", "pearson"))
}
\arguments{
\item{fit1}{first fit from rowcorMatVec}

\item{fit2}{second fit from rowcorMatVec}
}
\value{
rowcorMatVec object, slightly modified
}
\description{
when pearson and spearman values both are desired, we can fit them separately and then combine
}

