% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{corTest}
\alias{corTest}
\title{correlation betweent two vector}
\usage{
corTest(x, y, method = "spearman", minN = 5, ...)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{method}{method use}
}
\value{
a vector
}
\description{
correlation betweent two vector
}

