% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{df_force_factor}
\alias{df_force_factor}
\title{force selected columns of a data frame as factor}
\usage{
df_force_factor(dat, cols)
}
\arguments{
\item{dat}{a data frame}

\item{cols}{either an index vector or a string character for colnames to be applied with as.factor}
}
\description{
this is just a general (coarse) approach; if need to specify levels, do it manually.
}

