% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{df_format_df}
\alias{df_format_df}
\title{format a data frame by imposing non-continuous variables as factor though guessing}
\usage{
df_format_df(dat, minCat = 10)
}
\arguments{
\item{dat}{a data frame}

\item{dat}{a data frame}
}
\value{
a formated data frame
}
\description{
continuous variables are guessed by the is_continuous function
}
\details{
End product: all character/logical variables are forced to be factor; some integers or numeric would be forced to factor when they do not
satisfy minimum categories
}

