% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{dichotomizeExpr}
\alias{dichotomizeExpr}
\title{dichotomize expression into binary based on BI or quantile}
\usage{
dichotomizeExpr(vec, method = "quantile", q1 = NA, q2 = NA,
  Lower = NULL, Upper = NULL, labels = c("Low", "High"), plot = TRUE,
  forceFactor = FALSE)
}
\arguments{
\item{vec}{expressionv ector}

\item{method}{either quantile or BI}

\item{q1}{quantile 1 (smaller)}

\item{q2}{quantile 2 (Laerger) when this specified, it overrides Lower/Upper specification}

\item{labels}{what you want to call the two groups}
}
\value{
a character vector
}
\description{
This is a wrapper for dichotomizeExpr_quantile() and dichotomizeExpr_BI()
}

