% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{dichotomizeExpr_BI}
\alias{dichotomizeExpr_BI}
\title{dichotomize expression into binary based on BI}
\usage{
dichotomizeExpr_BI(vec, labels = c("Low", "High"), plot = FALSE)
}
\arguments{
\item{vec}{expressionv ector}

\item{labels}{what you want to call the two groups}
}
\value{
a character vector
}
\description{
This is usually needed for KM curve or building surragote resonse variable from continuous variable
}

