% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{dichotomizeExpr_quantile}
\alias{dichotomizeExpr_quantile}
\title{dichotomize expression into Low/High group}
\usage{
dichotomizeExpr_quantile(vec, q1 = NA, q2 = NA, Lower = NULL,
  Upper = NULL, labels = c("Low", "High"))
}
\arguments{
\item{vec}{expressionv ector}

\item{q1}{quantile 1 (smaller)}

\item{q2}{quantile 2 (Laerger) when this specified, it overrides Lower/Upper specification}

\item{labels}{what you want to call the two groups}
}
\value{
a character vector with middle values as NA
}
\description{
It's not recommeded to do this. But biologists wants to do it...
}
\details{
values exceeding Upper limit are called High; values smaller than Lower are called low. Values inbetween will be masked as NA.
Warning: When Lower and Upper are the same (when i.e. 40% not expressed, 1/3 and 2/3 quantiles are the same value!), only one group is made!
}

