% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_p.R
\name{doccoRmd}
\alias{doccoRmd}
\title{generate html from Rmd with docco-linear style as in \url{http://cran.r-project.org/web/packages/knitr/vignettes/docco-linear.html}}
\usage{
doccoRmd(input, ...)
}
\arguments{
\item{input}{input Rmd file}

\item{...}{additional parameter to knit2html (actually markdownToHTML)}
}
\description{
This generates kind of good looking html
}
\details{
Notice that envir is set to be globalenv() so that after this evaluation, all objects are still in the console.
Otherwise, some child script cannot run since it is not in the global.
}

