% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{fitSIBERonMat_Bygroup}
\alias{fitSIBERonMat_Bygroup}
\title{fit SIBER on matrix but split by group}
\usage{
fitSIBERonMat_Bygroup(mat, group, addAll = TRUE, addAllName = "all", ...)
}
\arguments{
\item{mat}{matrix of expression, methylation and so on}

\item{group}{a vector indicating the grouping membership}

\item{addAllName}{the name for the all group}

\item{...}{additional parameters to fitSIBERonMat such as model='NL', prune=TRUE, q=0.01, parallel=TRUE, core=10}
}
\value{
a data frame of BI
}
\description{
this is written to facilitate pancancer analysis where BI is computed across tumors, thus, a group variable
}

