% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{fn_TWanova}
\alias{fn_TWanova}
\title{two way anova test for a given response and two vectors}
\usage{
fn_TWanova(y, mainVec, confounderVec, interaction = TRUE,
  varLabel = c("Treatment", "Cell line", "Interaction"))
}
\arguments{
\item{interaction}{whether to add interaction term}

\item{varLabel}{the label for the 3 effects: main effect of primary variable, main effect of variable not of interest (confounding)
and interaction effect}
}
\value{
a fn_TWanova object with plot method
}
\description{
two way anova test for a given response and two vectors
}

