% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{formatLevel}
\alias{formatLevel}
\title{formatting a given vector according to given levels}
\usage{
formatLevel(vec, levels = NULL)
}
\arguments{
\item{vec}{a vector}

\item{levels}{a vector of levels specified for vec}
}
\value{
a factor
}
\description{
formatting a given vector according to given levels
}

