% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{get_pal}
\alias{get_pal}
\title{obtain color pallete defined in this package}
\usage{
get_pal(pal = NULL)
}
\arguments{
\item{pal}{the name for color pallete; if NULL, all will be returned in a list}
}
\description{
obtain color pallete defined in this package
}

