% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{get_stretch}
\alias{get_stretch}
\title{compute stretch from a continuous map}
\usage{
get_stretch(map, ref = 10)
}
\arguments{
\item{map}{a map}
}
\value{
a value
}
\description{
compute stretch from a continuous map
}

