% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{getfisherP}
\alias{getfisherP}
\title{compute fisher exact test (chisq test when failure) p value for two categorical variable}
\usage{
getfisherP(x, y, isBinary = FALSE)
}
\arguments{
\item{x}{categorical var 1}

\item{y}{categorical var 2}

\item{isBinary}{if both are binary variable 0, 1, we can compute OR to determine direction of association}
}
\description{
compute fisher exact test (chisq test when failure) p value for two categorical variable
}

