% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{getPvecSmry}
\alias{getPvecSmry}
\title{given a p value vector and a vector of cutoffs, find number of genes as well as indeces for genes passing these cutoffs}
\usage{
getPvecSmry(P, thr = c(0.001, 0.01, 0.05))
}
\arguments{
\item{P}{a vector of P values}

\item{thr}{a vector of thersholds}
}
\value{
a list with elements smryTab, selList, cutoff
}
\description{
given a p value vector and a vector of cutoffs, find number of genes as well as indeces for genes passing these cutoffs
}

