% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{getter.augMat}
\alias{getter.augMat}
\title{get method for class augMat}
\usage{
\method{getter}{augMat}(obj, what = "effectiveMat")
}
\arguments{
\item{obj}{an augMat object}

\item{what}{specify what to get from augMat, i.e. effectiveMat, isNAcol}
}
\value{
an object as requested
}
\description{
this function is named getter, not get to minimize conflict with the usual get function
}

