% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{getter.colpal4DF}
\alias{getter.colpal4DF}
\title{get method for class colpal4DF}
\usage{
\method{getter}{colpal4DF}(obj, what = "avail")
}
\arguments{
\item{obj}{an colpal4DF object}

\item{what}{specify what to get from augMat, i.e. base, var.equal}
}
\value{
an object as requested
}
\description{
this function is named getter, not get to minimize conflict with the usual get function
}

