% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{Getter}
\alias{Getter}
\title{general getter function for attributes in an object}
\usage{
Getter(obj, what = "avail")
}
\arguments{
\item{obj}{object with attributes}

\item{what}{specify what attribute to get}
}
\value{
attributes extracted
}
\description{
general getter function for attributes in an object
}

