% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{getter.rowanovaMatVec}
\alias{getter.rowanovaMatVec}
\title{get method for class rowanovaMatVec}
\usage{
\method{getter}{rowanovaMatVec}(obj, what = "avail")
}
\arguments{
\item{obj}{an rowanovaMatVec object}

\item{what}{specify what to get from augMat, i.e. base, var.equal}
}
\value{
an object as requested
}
\description{
this function is named getter, not get to minimize conflict with the usual get function
}

