% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{getter.rowttestMatVec}
\alias{getter.rowttestMatVec}
\title{get method for class rowttestMatVec}
\usage{
\method{getter}{rowttestMatVec}(obj, what = "meta")
}
\arguments{
\item{obj}{an rowttestMatVec object}

\item{what}{specify what to get from augMat, i.e. base, var.equal}
}
\value{
an object as requested
}
\description{
this function is named getter, not get to minimize conflict with the usual get function
}

