% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{getTopNvalue}
\alias{getTopNvalue}
\title{get top N elements from a vector}
\usage{
getTopNvalue(vec, N = NULL)
}
\arguments{
\item{vec}{a vector of values}

\item{N}{number of elemetns to select}
}
\value{
indeces corresponding to top N values from largest to smallest
}
\description{
get top N elements from a vector
}

