% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_survival_p.R
\name{ggKMcat}
\alias{ggKMcat}
\title{for categorical covariate x, plot KM curve}
\usage{
ggKMcat(x, OS, OScensoring, main, cex.main = 0.8, tag,
  xlab = "Survival time", ylab = "Survival probability", labels,
  legendTitle, nrowL = NA, sizeaxis = 12, sizeaxistitlerel = 1.2,
  sizetitlerel = 0.8, theme_bg = theme_base(), cols, legend_within = TRUE,
  lx = 0.8, ly = 0.88, plot = TRUE, verbose = FALSE,
  addEventCount = F, countFromInput = TRUE)
}
\arguments{
\item{x}{a factor}

\item{OS}{survival time}

\item{OScensoring}{censoring status, 1 is event}

\item{main}{title}

\item{tag}{a tag added at the start of title}

\item{xlab}{x label}

\item{ylab}{y label}

\item{labels}{this is deprecated since labels are just values from x}

\item{legendTitle}{legend title name}

\item{nrowL}{legend row number}

\item{theme_bg}{background theme}

\item{cols}{either a named vector (names from x) to accurately specify the color or a color vecto which is to be input in the alphabetical order as the x values}

\item{legend_within}{whether to put legend within the figure}

\item{lx}{relative x position of legend (between 0~1)}

\item{ly}{relative y position of legend (between 0~1)}

\item{verbose}{whether to print warning due to survframe count incompatible with observed.}

\item{addEventCount}{addEventCount whether to add event count in the legend}
}
\description{
This wraps qplot_survival() function for KM curve generation.
}
\details{
Note no sample size indicated in text: NA omitted in x, the strata var, however, OS can still be missing
so this is nominal N, effective N due to OS missing is not computed; this is good for tracking sample size and avoid confusion. We just
need to bear in mind that some samples will miss OS.

modified 2014/02/14: when x has only one unique value, get error: Error in `$<-.data.frame`(`*tmp*`, "strata", value = character(0)) : 
 replacement has 0 rows, data has 68
What color corresponds to what label? The labels by default are extracted from levels(x); The cols and computed labels (after
adding N) are 1-to-1 mapping. So it is a matter wheather the color is pointing to the same curve, which depends on qplot_survival() function.
I cannot prove qplot_survival() gets the color correct. But it seems work for categorical variable at this moment.
Updates: qplot_survival() computed sframe which is a df format for ggplot2 where a variable strata is formed which makes it possible
to use aes(group=strata, colour=strata). By default, the labels in strata is in alphabetical order;  
------------------->
Now, ggKMcat does not specify the labels. Instead, if one wants to change the labels, just rename the values in x through mapNames. 
Since the color and curves are created by aes(group, colour), there is always correct correspondance between curve and colour. The cols can be specified through a named
vector so that one can accurately control the color for each curve. 
To do: add N to the legend

To check with traditional surv curve, do:
		mysurv <- with(clin1u, survfit(Surv(PFS, PFScensoring) ~ GENDER_PHENO))
}

