% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_survival_p.R
\name{guess_nrowL}
\alias{guess_nrowL}
\title{guess nrowL from a specified number of categories}
\usage{
guess_nrowL(nCategory, useMax = FALSE)
}
\arguments{
\item{nCategory}{number of categories}

\item{useMax}{whether to use maximum value}
}
\description{
guess nrowL from a specified number of categories
}

