% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{guessAlpha_line}
\alias{guessAlpha_line}
\title{guess alpha in a line plot}
\usage{
guessAlpha_line(G)
}
\arguments{
\item{G}{number of genes or lines}
}
\value{
a scalar between 0 and 1
}
\description{
guess alpha in a line plot
}

