% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{guessCEX}
\alias{guessCEX}
\title{given number of genes G in a heatmap, guess rowCEX to be passed to cexRow}
\usage{
guessCEX(G)
}
\arguments{
\item{G}{number of genes}
}
\value{
a scalar between 0 and 1
}
\description{
given number of genes G in a heatmap, guess rowCEX to be passed to cexRow
}

