% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{heatm.rowcorMatVec}
\alias{heatm.rowcorMatVec}
\title{Heatmap plot for rowcorMatVec class}
\usage{
\method{heatm}{rowcorMatVec}(x, plot = TRUE, return = TRUE,
  gSelForce = NULL, gSel = "gSelectInd_P05", colbarname = "vec",
  revlog10 = FALSE, plotLegend = "TRUE", ...)
}
\arguments{
\item{x}{a rowcorMatVec class (a data frame)}

\item{gSel}{string specifying what genes are selected in the elements of 'obj_bum'}

\item{revlog10}{sometimes vec is log10 IC50; it is better to show IC50 in original scale in the heatmap; this will be possible by specifying revlog10=TRUE}
}
\description{
Heatmap plot for rowcorMatVec class
}

