% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{hist.rowfisherMatVec}
\alias{hist.rowfisherMatVec}
\title{Bum plot for rowfisherMatVec class}
\usage{
\method{hist}{rowfisherMatVec}(x, main = NULL, plot = TRUE,
  return = FALSE, alphas = c(0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5),
  FDR = NULL, maskPthr = 1, pcutoff = 0.05, ...)
}
\arguments{
\item{x}{a rowfisherMatVec class (a data frame)}

\item{...}{additional parameters to plotBumFDR}
}
\description{
Bum plot for rowfisherMatVec class
}

