% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{hist.rowttestMatVec}
\alias{hist.rowttestMatVec}
\title{Bum plot for rowttestMatVec class}
\usage{
\method{hist}{rowttestMatVec}(x, main = NULL, tag = "", plot = TRUE,
  return = FALSE, alphas = c(0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5),
  FDR = NULL, maskPthr = 1, pcutoff = 0.05, ...)
}
\arguments{
\item{x}{a rowttestMatVec class (a data frame)}

\item{...}{additional parameters to plotBumFDR}
}
\description{
Bum plot for rowttestMatVec class
}

