% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{hist.rowTWanovaMatVec}
\alias{hist.rowTWanovaMatVec}
\title{Bum plot for rowTWanovaMatVec class}
\usage{
\method{hist}{rowTWanovaMatVec}(x, var = "pval_main",
  varLabel = "Treatment effect p value", main = NULL, plot = TRUE,
  return = FALSE, alphas = c(0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5),
  FDR = NULL, maskPthr = 1, pcutoff = 0.05)
}
\arguments{
\item{x}{a rowTWanovaMatVec class (a data frame)}

\item{var}{what variable (pval_main, pval_confounder, pval_interaction, anovaP|HCC1806, ...) for bum plot; if NULL, all are plotted}

\item{varLabel}{label to be used for annotation purpose describing var}

\item{...}{additional parameters to plotBumFDR}
}
\description{
Bum plot for rowTWanovaMatVec class
}

