% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{is_continuous}
\alias{is_continuous}
\title{guess if a vector is continuous}
\usage{
is_continuous(vec, minCat = 10)
}
\arguments{
\item{vec}{a vector}

\item{minCat}{an integer specifying minimum categories required to be a continuous variable for integers}
}
\value{
TRUE for continuous and FALSE for anything else
}
\description{
we treat character as categorial; double as continuous; for integer/numeric/double, we can specify a minimum categories to be continuous variable
}
\details{
a vector of class character is for sure categorical and would be good to handled as a factor in R. The class numeric
can be either integer (is.integer) or double (is.double). If is.double, then we should treat it as continuous when it has enough categories (binary
can be stored as double mode!).
For is.integer, when there is less than 5 categories, i.e., it is better treated as categorical.
}

