% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{iscat}
\alias{iscat}
\title{check if a vector is categorical or continuous}
\usage{
iscat(vec)
}
\arguments{
\item{vec}{a vector}
}
\value{
logical
}
\description{
check if a vector is categorical or continuous
}

