% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{iscatDF}
\alias{iscatDF}
\title{check columnwise of a data frame whether it is categorical (character or factor) or continuous}
\usage{
iscatDF(dat)
}
\arguments{
\item{dat}{a data frame}
}
\value{
logic vector
}
\description{
this is a colwise application of iscat()
}

