% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{isTopN}
\alias{isTopN}
\title{return a logical vector if elements in a vector is top N largest}
\usage{
isTopN(vec, N = 100)
}
\arguments{
\item{vec}{a numeric vector}

\item{N}{top N}
}
\value{
logic vector
}
\description{
return a logical vector if elements in a vector is top N largest
}

