% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{ixglobal2local}
\alias{ixglobal2local}
\title{map global index to local index given reference}
\usage{
ixglobal2local(global, ref)
}
\arguments{
\item{global}{global index}

\item{ref}{ref index based on global index}
}
\value{
index based on local reference
}
\description{
map global index to local index given reference
}

