% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{ixlocal2global}
\alias{ixlocal2global}
\title{map local index to global index given reference}
\usage{
ixlocal2global(local, ref)
}
\arguments{
\item{local}{local index based on ref}

\item{ref}{ref index based on global index}
}
\value{
index based on global as reference
}
\description{
map local index to global index given reference
}

