% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{logic2sortedIndex}
\alias{logic2sortedIndex}
\title{conver logic vector to sorted index vector}
\usage{
logic2sortedIndex(logicVec, valVec, decreasing = TRUE)
}
\arguments{
\item{logicVec}{logic vector}

\item{valVec}{value vector; should have 1-to-1 correspondence to logicVec}

\item{decreasing}{used in sorting}
}
\value{
a vector of index
}
\description{
given a logic vector and corresponding value for ordering, reuturn an index vector where logicVec is TRUE and sort the index by valVec
}

