% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{maskByLimit}
\alias{maskByLimit}
\title{mask values as NA in a matrix}
\usage{
maskByLimit(x, Lower, Upper)
}
\arguments{
\item{x}{a matrix}

\item{Lower}{lower cutoff value}

\item{Upper}{upper cutoff value}
}
\value{
formated matrix
}
\description{
mask values as NA in a matrix
}

