% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplots_p.R
\name{mylibrary}
\alias{mylibrary}
\title{multiple scatter plot for a given data frame}
\usage{
mylibrary(..., mute = FALSE)
}
\arguments{
\item{...}{a character string, no need to use c()}

\item{mute}{whether to mute}
}
\description{
not implemented due to technical difficulty in passing arguments in a nested function wrapper!
smart library function
}
\details{
load a library; when not present, install it
}

