% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{n_unique}
\alias{n_unique}
\title{return the number of unique values}
\usage{
n_unique(x, ignoreNA = FALSE)
}
\arguments{
\item{x}{a vector}

\item{ignoreNA}{logical, whether to count NA as a separate unique value; default is count NA as one unique value}
}
\value{
integer
}
\description{
return the number of unique values
}

