% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{name of the vec as in the named list of colpal}
\alias{name of the vec as in the named list of colpal}
\alias{vecpal2map}
\title{build a map given a vector of values and color pallete}
\usage{
vecpal2map(vec, pal, name, iscat, n = 59)
}
\arguments{
\item{vec}{a vector of values to be mapped}

\item{pal}{color pallete}

\item{iscat}{logical indicating if cont or cat}

\item{n}{only used to specify number of colors generated for continious vec}
}
\value{
a color map
}
\description{
for continuous variable, makecmap is used; for categorical variable, buildcolorfacs is used
}

