% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{noNA}
\alias{noNA}
\title{remove NA cases for either a vector, matrix or data frame}
\usage{
noNA(dat, returnIndex = FALSE)
}
\arguments{
\item{returnIndex}{whether just to return index of rows}

\item{data}{a vector, data frame or matrix}
}
\value{
same format as dat
}
\description{
obtain complete records from your data
}

