% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{nonNumeric2FactorByColumn}
\alias{nonNumeric2FactorByColumn}
\title{force to factor when non-numeric}
\usage{
nonNumeric2FactorByColumn(df)
}
\arguments{
\item{df}{it is better to be a data frame. if a vector or matrix, it will be internally converted to data.frame}
}
\value{
a data frame
}
\description{
for a given data frame, this function converts non-numeric columns into factor while the numeric columns are intact.
}

