% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_p.R
\name{overwritef}
\alias{overwritef}
\title{a wrapper for file overwrite}
\usage{
overwritef(FUN, file, overwrite = TRUE, ...)
}
\arguments{
\item{FUN}{a function to pass}

\item{file}{a file including path}

\item{overwrite}{TRUE or FALSE}

\item{...}{additional parameters for FUN}
}
\value{
Ord indicator if there is overwriting happen so that further code can be conditionally run, i.e. plot
}
\description{
typical usage is to pass any function that has a file argument, i.e. save(), pdf() and so on
}
\details{
this allows us to decide whether to overwrite existing file in a directory.
}
\examples{
pdfOR(file='test.pdf', overwrite=TRUE)
plot(rnorm(100))
dev.off()
overwritef(pdf, file='test.pdf', overwrite=TRUE)
plot(1:3)
dev.off()
}

