% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{parse_coxphP2}
\alias{parse_coxphP2}
\title{parse the coxphP2() returned value to text so as to show OR, P and so on}
\usage{
parse_coxphP2(x, type = "HR", digits = 2)
}
\arguments{
\item{x}{results returned by coxphP2()}

\item{type}{either OR from categorical variable or HR from continuous variable}

\item{digits}{digits to keep in OR and CI}
}
\value{
a list of text string. i.e.

$coefExp_text

[1] "OR: 1.41"

$CI_text

[1] "CI: [0.88, 2.27]"

parse coxphP2 result
}
\description{
parse the coxphP2() returned value to text so as to show OR, P and so on
}

