% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_p.R
\name{pdfOR}
\alias{pdfOR}
\title{pdf() with overwrite option}
\usage{
pdfOR(file, overwrite = TRUE, ..., return = TRUE)
}
\arguments{
\item{return}{default to have return}
}
\description{
pdf() with overwrite option
}

