% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{plot.colpal4DF}
\alias{plot.colpal4DF}
\title{plot method for colpal4DF}
\usage{
\method{plot}{colpal4DF}(obj, nr = 12, mar = c(2, 9, 2, 1))
}
\arguments{
\item{obj}{a colpal4DF class usually returned by autocolpal4DF()}

\item{nr}{number of colors to show}

\item{mar}{mar in par() setting}
}
\description{
plot method for colpal4DF
}

