% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{plot.fn_TWanova}
\alias{plot.fn_TWanova}
\title{Plot method for fits from two-way data by fn_TWanova}
\usage{
\method{plot}{fn_TWanova}(x, pshape = 17, psize = 3.5, tag = "",
  cols = NULL, main = NULL, xlab = NULL, ylab = "Expression",
  legendLabelSize = 11, anglex = NULL, sizex = 12, sizey = 12,
  angley = 0, plot = T)
}
\arguments{
\item{cols}{a named color vector where names are the unique group categories. Default color is specified with sorted group category.}

\item{legendLabelSize}{legend label size}
}
\description{
This in the title shows the p values for main effects (two variables namely treatment and confounder) and interaction
The figure is boxplot colored by confounder; at the bottome, stratified anova p value is showing (for each color)
}

