% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{plot.phm}
\alias{plot.phm}
\title{plot method for phm class}
\usage{
\method{plot}{phm}(obj, y0 = 1, x0 = 0.8, yd = 0.17)
}
\arguments{
\item{obj}{a phm class (a data frame)}

\item{...}{additional parameters to plot4rowanovaMatVec}
}
\description{
plot method for phm class
}

