% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{plot.prepcolbar}
\alias{plot.prepcolbar}
\title{plot method for prepcolbar class}
\usage{
\method{plot}{prepcolbar}(obj, cex.row = NULL, cex.col = NULL, mar = c(2,
  9, 2, 1))
}
\arguments{
\item{obj}{a prepcolbar class (a data frame)}

\item{cex.row}{cex.row}

\item{cex.col}{cex.col}

\item{mar}{margin setting}
}
\description{
plot method for prepcolbar class
}

