% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{plot.rowanovaMatVec}
\alias{plot.rowanovaMatVec}
\title{Bar plot for rowanovaMatVec class}
\usage{
\method{plot}{rowanovaMatVec}(x, maxRow = 400, main = NULL,
  dataMode = NULL, groupShown = NULL, ...)
}
\arguments{
\item{x}{a rowanovaMatVec class (a data frame)}

\item{...}{additional parameters to plot4rowanovaMatVec}
}
\description{
Bar plot for rowanovaMatVec class
}

