% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{plot.rowcorMatVec}
\alias{plot.rowcorMatVec}
\title{Bar plot for rowcorMatVec class}
\usage{
\method{plot}{rowcorMatVec}(x, maxRow = 400, main = NULL, tag = "", ...)
}
\arguments{
\item{x}{a rowcorMatVec class (a data frame)}

\item{...}{additional parameters to plot4rowttestMatVec}
}
\description{
Bar plot for rowcorMatVec class
}

