% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{plot.rowttestMatVec}
\alias{plot.rowttestMatVec}
\title{Bar plot for rowttestMatVec class}
\usage{
\method{plot}{rowttestMatVec}(x, maxRow = 400, main = NULL, tag = "",
  dataMode = NULL, fillCol = c("#ED281E", "#2834A4", "#EDB51E", "gray"),
  ...)
}
\arguments{
\item{x}{a rowttestMatVec class (a data frame)}

\item{fillCol}{color for bar filling in the order of c('P<0.001', 'P<0.01', 'P<0.05', 'P>0.05')}

\item{...}{additional parameters to plot4rowttestMatVec}
}
\description{
Bar plot for rowttestMatVec class
}

