% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{plot4rowanovaMatVec}
\alias{plot4rowanovaMatVec}
\title{ordered horizontal bar plot for rowanovaMatVec}
\usage{
plot4rowanovaMatVec(x, rowInfo = NULL, decreasing = FALSE,
  dataMode = "IC50", groupShown = NULL, ylab = NULL, xlab = NULL,
  sizex = 12, sizey, size_group = 10, main = NULL, showP = "None",
  pthr = 1, colorSignif = TRUE, legend.position = c(0.85, 0.2),
  legend.title = "Significance", truncFClow = -Inf, truncFChigh = Inf,
  barWidth = 0.8, verbose = FALSE)
}
\arguments{
\item{rowInfo}{alternative rownames to show on the figure}

\item{decreasing}{logic, decreasing order of p value used to sort the rows (genes) based on anovaP.}

\item{dataMode}{a string used to label the x-axis. Default is 'IC50' for IC50 t-test; Please use meaning string relevant to the data.}

\item{groupShown}{string for group to shown, leading a t test type (FC vs adj p value). e.g. 'HCC4006_ER1-HCC4006'}

\item{sizex}{x axis label size}

\item{sizey}{y axis label size}

\item{size_group}{label size for facets}

\item{showP}{do not show P value on graph ('None'); show actual P ('Value'); show P category ('Category')}

\item{pthr}{filtering by p value. Rows with larger P value will be deleted. Default is to use all rows (pthr=1)}

\item{legend.position}{position for the legend if P value is show as category}

\item{legend.title}{title of the legend}

\item{truncFClow}{cutoff to truncate FC lower than this value. Defalt is -Inf, which means no truncation on small FC (negative) value}

\item{truncFChigh}{cutoff to truncate FC higher than this value. Defalt is Inf, which means no truncation on large FC (usually positive) value}

\item{barWidth}{relative barwidth (0~1)}
}
\value{
a ggplot2 object
}
\description{
ordered horizontal bar plot for rowanovaMatVec
}

