% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{plot4rowcorMatVec}
\alias{plot4rowcorMatVec}
\title{ordered bar plot for rowcorMatVec class}
\usage{
plot4rowcorMatVec(res, rowInfo = NULL, decreasing = TRUE,
  dataMode = "IC50", ylab = NULL, xlab = NULL, sizex, sizey = 12,
  main = NULL, tag = "", showP = "None", pthr = 1, colorSignif = TRUE,
  legend.position = c(0.85, 0.85), legend.title = "Significance",
  showLegend = TRUE, truncCorlow = -1, truncCorhigh = 1,
  barWidth = NULL, corlim, coord_flip = FALSE, plot = TRUE)
}
\arguments{
\item{rowInfo}{alternative rownames to show on the figure}

\item{decreasing}{logic, decreasing order of bars or not.}

\item{dataMode}{a string used to label the x-axis. Default is 'IC50' for IC50 t-test; Please use meaning string relevant to the data.}

\item{sizex}{x axis label size}

\item{sizey}{y axis label size}

\item{showP}{do not show P value on graph (as part of gene name) ('None'); show actual P ('Value'); show P category ('Category')}

\item{pthr}{filtering by p value. Rows with larger P value will be deleted. Default is to use all rows (pthr=1)}

\item{legend.position}{position for the legend if P value is show as category}

\item{legend.title}{title of the legend}

\item{truncCorlow}{cutoff to truncate correlation lower than this value. Defalt is -1, which means no truncation on small corr (negative) value}

\item{truncCorhigh}{cutoff to truncate correlation higher than this value. Defalt is 1, which means no truncation on large corr (usually positive) value}

\item{barWidth}{relative barwidth (0~1); if this is too small, some bars might be displayed as missing; so set it larger}

\item{corlim}{correlation limit; default is not specified which will use observed correlation limit; can specify c(-1, 1) to see this range.}

\item{coord_flip}{whether to flip coordinate; default sets y as the correlation; when coord_flip=TRUE, x is plotted as correlation}
}
\value{
a ggplot2 object
}
\description{
ordered bar plot for rowcorMatVec class
}

