% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{plot4rowttestMatVec}
\alias{plot4rowttestMatVec}
\title{ordered horizontal bar plot for rowttestMatVec}
\usage{
plot4rowttestMatVec(res, useFC = FALSE, rowInfo = NULL, decreasing = TRUE,
  dataMode = "IC50", ylab = NULL, xlab = NULL, sizex = 12, sizey,
  sizeyadj = 1, size.title.x = 14, size.title.y = 14, main = NULL,
  showP = "None", pthr = 1, colorSignif = TRUE,
  legend.position = c(0.85, 0.85), legend.title = "Significance",
  truncFClow = -4, truncFChigh = 4, barWidth = 0.8, verbose = FALSE,
  fillCol = c("#ED281E", "#2834A4", "#EDB51E", "gray"), coord_flip = TRUE,
  anglex = NULL, angley = NULL, plot = TRUE)
}
\arguments{
\item{useFC}{whether to use FC or mDiff}

\item{rowInfo}{alternative rownames to show on the figure}

\item{decreasing}{logic, decreasing order of bars or not.}

\item{dataMode}{a string used to label the x-axis. Default is 'IC50' for IC50 t-test; Please use meaning string relevant to the data.}

\item{sizex}{x axis tick text size}

\item{sizey}{y axis tick text size}

\item{size.title.x}{x label size}

\item{size.title.y}{y label size}

\item{showP}{do not show P value on graph ('None'); show actual P ('Value'); show P category ('Category')}

\item{pthr}{filtering by p value. Rows with larger P value will be deleted. Default is to use all rows (pthr=1)}

\item{legend.position}{position for the legend if P value is show as category}

\item{legend.title}{title of the legend}

\item{truncFClow}{cutoff to truncate FC lower than this value. Defalt is -Inf, which means no truncation on small FC (negative) value}

\item{truncFChigh}{cutoff to truncate FC higher than this value. Defalt is Inf, which means no truncation on large FC (usually positive) value}

\item{barWidth}{relative barwidth (0~1)}
}
\value{
a ggplot2 object
}
\description{
ordered horizontal bar plot for rowttestMatVec
}

