% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplots_p.R
\name{plotCatCat}
\alias{plotCatCat}
\title{plot cat-cat}
\usage{
plotCatCat(...)
}
\arguments{
\item{x}{a factor}

\item{y}{a factor}

\item{xlab}{xlab}

\item{ylab}{ylab}
}
\description{
a wrappter to the plotHeatmapTable function
TODO: add options for type=c('barprop', 'barN', 'piefacet', 'residual')
}

