% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplots_p.R
\name{plotContCont}
\alias{plotContCont}
\title{plot Cont-Cont variable}
\usage{
plotContCont(x, y, main = NULL, tag = "", xlab = "", ylab = "",
  theme_bw = TRUE, alpha = 1, size = 5, sizeExpand = 6, shape = 20,
  sizeaxis = 12, sizeaxistitlerel = 1.2, sizetitlerel = 0.8, lm = FALSE,
  lmcol = "blue2", lmsize = 2, lmlty = 2, smooth = FALSE, plot = TRUE,
  ylim = NULL, xlim = NULL, labels = NULL, sizelab = 3,
  colab = "purple", vjust = 1.9, hjust = 0.5, expand = NULL,
  plot.margin.cm = c(0.2, 1.2, 0.2, 0.2))
}
\arguments{
\item{x}{continuous vector}

\item{y}{continuous vector}

\item{tag}{used to add a tag to the title}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{theme_bw}{logical if bw theme is imposed}

\item{alpha}{alpha transparancy for the points}

\item{size}{size passed to geom_point()}

\item{sizeExpand}{size when expanded}

\item{shape}{shape passed to geom_point()}

\item{sizeaxis}{axis text size (tick text)}

\item{sizeaxistitlerel}{relative axis title size (tick text)}

\item{sizetitlerel}{relative size for main title}

\item{lm}{if to add a lm fit line}

\item{smooth}{if to add a smooth line}

\item{ylim}{ylim}

\item{xlim}{xlim}

\item{labels}{a vector of IDs to be added to each data point}

\item{sizelab}{size for label under each data point}

\item{expand}{whether to expand the size and axis title (useful if figure is to be copied in slides with multiple panels)}

\item{plot.margin.cm}{the right margin might be too small to show full axis text; this specifies margins in top, right, bottom, left}
}
\description{
Scatter plot for continuous variables and associated statistics
}
\details{
for cont~cont, we need a scatter plot. An lm fit with p value and R^2. A trend fit with smoothing can be added. Spearman corr. and Pearson Corr.
}

