% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{present a heatmap}
\usage{
plotHeatmap(smat_tx, colbar, rowbar = NULL,
  color = colorpalette2colvec("blueyellow"), ncolor = 60,
  colBarSel = NULL, oma = c(2, 1, 3, 11) + 0.1, plot = TRUE,
  cexRow = NULL, cexCol = NULL, labCol = NULL, labRow = NULL,
  Kcol = NULL, Krow = NULL, labColpal = "set1", labRowpal = "set2",
  labRowcolor = NULL, ...)
}
\arguments{
\item{colbar}{column bar object as returned by prepcolbar(): need to add option for colbar=NULL, no colbar case}

\item{rowbar}{row bar object as returned by prepcolbar(). Default is NULL, meaning no rowbar}

\item{color}{color vector for expression data}

\item{ncolor}{number of colors to be interpolated based on color parameter}

\item{colBarSel}{a vector to select a subset of column bars; this should be a subset of colnames of pheno data (also colnames of colbar)}

\item{oma}{oma passed to par()}

\item{plot}{whether to plot the heatmap}

\item{...}{additional parameters to heatmat}
}
\value{
mapL list with additional attributes to reconstruct heatmap
 to reconstruct, use:
 with(attributes(phm), ...)
}
\description{
Constructing a heatmap can be divided into 3 steps:
 (1) prepare column bars using prepcolbar() function: specify colorbar color schemes
 (2) subsetting main matrix for gene selection and sample selection using subsetMainMat()
 (3) transform main matrix including scaling, gene/sample clustering, and ordering. this step follows (2) with smat 
     and can be done with transform_smat()
 (4) present heatmap (specify main heatmap colors)
}

