% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplots_p.R
\name{plotHeatmapTable}
\alias{plotHeatmapTable}
\title{plot a table with colour indicated}
\usage{
plotHeatmapTable(x, y, xlab, ylab, tag = NULL, main = NULL,
  verbose = TRUE, OrdReg = TRUE, plot = TRUE, plotType = "stdres",
  low.color = "white", high.color = "blue", text.size = 8,
  stdres_range = c(-4, 4))
}
\arguments{
\item{x}{a factor}

\item{y}{a factor}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{tag}{the tag on the title; a fisher exact test p value will be appended}

\item{verbose}{logical indicating whether print model fitting result on screen}

\item{OrdReg}{logical indicating if an ordinal regression is to be fitted}

\item{plot}{logical indicating if a plot is to be generated}

\item{plotType}{type of plot. stdres for coloring by std residuals. count for coloring by count. dull for no coloring}

\item{low.color}{the color for smallest value}

\item{high.color}{the color for largest value}

\item{text.size}{text size for the count}
}
\description{
Given a table object, this function plot the count and indicate the actual count
}
\details{
modified on 2014/01/29: fisher test might fail; add try-catch (i.e. user chisq or NA); default xlab is related to tab dimnames now.
modified on 2014/01/30: add ordinal logistic regression info
modified on 2014/11/26: when chisq is successfully fitted, cell will be colored by standardized residual to understand which cell is enriched
}
\examples{
x <- sample(LETTERS[1:4], 100, replace=TRUE)
y <- sample(LETTERS[5:8], 100, replace=TRUE)
}

