% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qplot_survival_p.R
\name{plotKM0}
\alias{plotKM0}
\title{plot just KM}
\usage{
plotKM0(OS, OScensoring, f.CI = TRUE, xlab = "Survival time", f.shape = 3,
  ylab = "Survival Probability", main = "Kaplan-Meier Curve",
  hcolor = "#a2abb5", legendText = "Death status", tag = "", lx = 0.2,
  ly = c(0.1, 0.05), showAnno = TRUE, cex.main = 0.8)
}
\arguments{
\item{OS}{Overall survival}

\item{OScensoring}{censoring status}

\item{f.CI}{whether to add C.I.}

\item{xlab}{x label}

\item{f.shape}{shae}

\item{ylab}{y label}

\item{main}{main title}

\item{hcolor}{horizontal line color}

\item{legendText}{text for legend}

\item{lx}{x position of legend}

\item{ly}{y position of legend}

\item{showAnno}{whether to show annotation}
}
\description{
no covariate KM plot
}

