% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplots_p.R
\name{plotMeanSD4mat}
\alias{plotMeanSD4mat}
\title{plot mean vs sd for a matrix for quick QC/filtering}
\usage{
plotMeanSD4mat(mat, type = "row", useMAD = FALSE, alpha = NA,
  qx = c(0.3, 0.5), qy = c(0.3, 0.5), main = NA,
  quantileCol = "#B0AEA9", plot = TRUE)
}
\arguments{
\item{mat}{a matrix}

\item{type}{compute summary on rows or columns}

\item{useMAD}{whether to use MAD (more robust) rather than SD. Default is FALSE}

\item{alpha}{alpha blending for the points}

\item{qx}{quantile to be show for Mean (0~1)}

\item{qy}{quantile to be show for SD (0~1)}

\item{main}{main title}

\item{quantileCol}{line colors for quantiles}

\item{plot}{plot or return the summary}
}
\description{
plot mean vs sd for a matrix for quick QC/filtering
}

