% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplots_p.R
\name{plotScatter}
\alias{plotScatter}
\title{flexible scatter plot (a wrapper)}
\usage{
plotScatter(...)
}
\arguments{
\item{...}{other parameters passed to either plotContCont plotContCat plotCatCat}

\item{x}{either numeric or factor or numeric}

\item{y}{either numeric or factor or numeric}
}
\description{
scatter for a cont-cont, cont-cat or cat-cat variable pair.
}
\details{
the program detects x, y as categorical if it is a factor. Then it wraps plotContCont plotContCat plotCatCat
}

