% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{polrP}
\alias{polrP}
\title{ordinal regression for association between an ordered categorical variable with a conitnuous variable}
\usage{
polrP(expr, label, addANOVA = TRUE, base = 2, more = TRUE)
}
\arguments{
\item{expr}{a vector of continuous values}

\item{label}{a categorical vector or factor}

\item{addANOVA}{whether to perform an ANOVA analysis}

\item{base}{additional parameters passed to anovaP; only effective if addANOVA is TRUE.}

\item{more}{additional parameters passed to anovaP; only effective if addANOVA is TRUE.}
}
\value{
a vector
}
\description{
ordinal regression for association between an ordered categorical variable with a conitnuous variable
}

