% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{prepcolbar}
\alias{prepcolbar}
\title{build colmat/colmap given pheno (colwise of data frame) and colpal}
\usage{
prepcolbar(pheno, colpal = NULL)
}
\arguments{
\item{pheno}{a data frame where columns represent different column bars (in a heatmap)}

\item{colpal}{color pallete, a named list of color vectors. When not specified, autocolpal4DF will be used to generate the default.}
}
\value{
a list with elements mapL (a list of map for color legend) and colmat (color matrix for colbar). 
 input pheno and colpal are attached as attributes
}
\description{
building colmap depends on iscat status of each column, catstat. catstat is first searched with attributes; if it is not specified, iscatDF is used to calculate it. 
therefore, the default assumes the user has prepared pheno in a way that categorical vars are either character or factor and cont vars are numeric.
}

