% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{print.phm}
\alias{print.phm}
\title{print method for phm class}
\usage{
\method{print}{phm}(obj)
}
\arguments{
\item{obj}{a phm class (a data frame)}

\item{...}{additional parameters to plot4rowanovaMatVec}
}
\description{
print method for phm class
}

