% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{rowanovaMatVec}
\alias{rowanovaMatVec}
\title{ANOVA between rows of a matrix and a vector}
\usage{
rowanovaMatVec(mat, vec, matIsX = TRUE, levels = NULL, meta = list(tag =
  "ANOVA", row = "expression", response = "response"), base = 2,
  more = TRUE, addData = TRUE, ...)
}
\arguments{
\item{mat}{a matrix, i.e. gene expression}

\item{vec}{a vector of phenotype, categorical variable}

\item{levels}{levels of the vec; when vec is not a factor, levels would be alphabetical order which will determine the scatter plot.}

\item{meta}{meta information, a list. This can specify what type of analysis (i.e. RPPA correlated with EMT score);
row, which represents what row variables are (rows of mat); response represents what does vec mean}
}
\value{
an object belonging to data frame and a class
}
\description{
ANOVA between rows of a matrix and a vector
}

