% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{rowcorMatMat_L}
\alias{rowcorMatMat_L}
\title{extension of rowcorMatVec}
\usage{
rowcorMatMat_L(mat, matR, method = "spearman", minN = 5, usercorr = TRUE,
  addData = TRUE, symbol = NULL, sepSymProbe = ", ",
  meta = list(responseSuffix = "", row = "", tag = "correlation"))
}
\arguments{
\item{mat}{a matrix, i.e. gene expression}

\item{matR}{a matrix for response; each row is a response; ncol should match with ncol(mat)}

\item{method}{method passed to correlation calculation. It can be spearman or pearson}

\item{minN}{minimum sample size for correlation calculation; If not achieved, correlation is returned as NA}
}
\description{
extension of rowcorMatVec
}

