% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{rowcorMatVec}
\alias{rowcorMatVec}
\title{Calculate correlation between rows of a matrix and a vector}
\usage{
rowcorMatVec(mat, vec, meta = list(tag = "correlation", row = "expression",
  response = "response"), method = "spearman", minN = 5, usercorr = TRUE,
  addData = TRUE, symbol = NULL, sepSymProbe = ", ", FDR = NULL,
  pcutoff = 0.05, alphas = c(0.01, 0.05, 0.1, 0.2, 0.3, 0.4, 0.5),
  addBum = NULL)
}
\arguments{
\item{mat}{a matrix, i.e. gene expression}

\item{vec}{a vector of phenotype, continuous variable}

\item{meta}{meta information, a list. This can specify what type of analysis (i.e. RPPA correlated with EMT score);
row, which represents what row variables are (rows of mat); response represents what does vec mean}

\item{method}{method passed to correlation calculation. It can be spearman or pearson}

\item{minN}{minimum sample size for correlation calculation; If not achieved, correlation is returned as NA}

\item{usercorr}{whether to use rcorr package for speeded calculation (using Fortran)}

\item{symbol}{in case this is probe level data, user can specify a vector of symbols matching}

\item{sepSymProbe}{separator for symbol-probe. That is, symbol+sepSymProbe+probe(rowname) is the new rowname}

\item{FDR}{FDR to create bum for the p values, passed to create_bum()}

\item{pcutoff}{pcutoff passed to create_bum()}

\item{alphas}{alphas passed to create_bum()}
}
\value{
a matrix with two columns: cor and pval
}
\description{
Calculate correlation between rows of a matrix and a vector
}

