% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{rowcorMatVec_Bygroup}
\alias{rowcorMatVec_Bygroup}
\title{the grouped version for rowcorMatVec}
\usage{
rowcorMatVec_Bygroup(mat, vec, group, addAll = FALSE, addAllName = "all",
  parallel = FALSE, core = 1, method = "spearman", minN = 5,
  usercorr = TRUE, returnArray = TRUE, symbol = NULL,
  sepSymProbe = ", ", ...)
}
\arguments{
\item{mat}{a matrix, i.e. gene expression}

\item{vec}{a vector of phenotype, continuous variable}

\item{group}{a vector, categorical variable}

\item{addAll}{whether to calculate correlation by using all data ignoring group; this is the overall correlation}

\item{parallel}{deprecated; parallel is automatically enabled when core>1}

\item{core}{number of cores to use; when core>1, parallel execution is enabled}

\item{...}{additional parameters to rowcorMatVec}

\item{addData}{logical indicating if data (mat) should be attached as an attribute (may be used for scatter plot)}
}
\value{
a 3D array
}
\description{
This calculated the correlation between rows of mat and vec but at each group at a time as specified by the group vector
It should be more flexible than rowcorMatMat which requires the same genomic mat; instead, here different group of analysis can
have different sample size in genomics
}

