% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{rowfisherMatVec}
\alias{rowfisherMatVec}
\title{Perform Fisher Exact test between rows of mat (categorical) and a factor}
\usage{
rowfisherMatVec(mat, vec, meta = list(tag = NULL, row = "group", response =
  "response"), more = TRUE, levels = NULL, parallel = FALSE, core = 1,
  addData = TRUE, ...)
}
\arguments{
\item{mat}{a matrix with rows as genes}

\item{vec}{a factor giving the response}

\item{meta}{meta information, a list. This can specify what type of analysis (i.e. KRAS vs NF1 mutation);
row, which represents what row variables are (rows of mat); response represents what does OS mean}

\item{more}{not implemented}

\item{levels}{levels specified for the response vec; rows of mat are formatted with as.factor currently; (this is not tested if it controls levels for plotting although it does not affect p value calculation)}

\item{core}{number of cores to use; this override parallel=TRUE if core=1 (no parallel)}

\item{addData}{logical indicating if data (mat) should be attached as an attribute (may be used for scatter plot)}

\item{...}{additional parameters to plotKM}
}
\value{
a rowfisherMatVec class; meta information is stored as attributes
}
\description{
This will perform fisher exact test; if this fails (e.g. due to large dimention or counts), chisq test will be used to get the result.
Originally, ggplot2 figures are attached; this may takes a long time to cache the objects in knitr as well as saving it. Now ggplot2 objects are not
computed and thus not returned to improve efficiency
}

