% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{rowlmMatVec}
\alias{rowlmMatVec}
\title{Fit a linear model with rows of mat and vec, possibly adding an adjustment variable (adj)}
\usage{
rowlmMatVec(mat, vec, adj, matIsX = TRUE, collapse = TRUE,
  parallel = FALSE, core = 12)
}
\arguments{
\item{mat}{a matrix}

\item{vec}{a vector, length of ncol(mat)}

\item{adj}{a vector for adjustment variable}

\item{matIsX}{by default, mat rows are continuous (i.e. expr) and vec is categorical response; but for CN data where mat is categorical and the response might be continuous or not.
in this case, we want lm(y~x) treats rows of mat as y and vec as x. This can be achieved by specifying matIsX=FALSE that makes rows of mat as the response variable}

\item{collapse}{when vec is non-numeric which will be treated as factor, there are multiple coefs for different levels. We thus need to
summarize them into a scaler (by mean). if collapse=FALSE, all coefs will be returned.}

\item{core}{number of cores to use for parallel}

\item{adj}{this can be a vector or a data frame that contains multiple variables to be adjusted. It needs to have nrow=length(vec).----> now demands a 1 column for easy handling}
}
\value{
a data frame
}
\description{
Fit a linear model with rows of mat and vec, possibly adding an adjustment variable (adj)
}

