% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{rowranktestMatVec}
\alias{rowranktestMatVec}
\title{compute rank test P as well as a t test result}
\usage{
rowranktestMatVec(mat, vec, core = 12, ...)
}
\arguments{
\item{mat}{a matrix with rows as genes}

\item{vec}{a vector giving the group info}

\item{core}{number of cores to use}

\item{...}{additional pars for rowranktestMatVec}
}
\value{
a data frame
}
\description{
this uses the wilcoxP() and ttestP(); the interface is similar to rowranktestMatVec.
}

