% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{rowTWanovaMatVec}
\alias{rowTWanovaMatVec}
\title{Two-way ANOVA between rows of a matrix and two vectors: two main effects and then interaction effect
 
Each row is to be fitted by fn_TWanova()}
\usage{
rowTWanovaMatVec(mat, vec1, vec2, matIsX = TRUE, interaction = TRUE,
  meta = list(tag = "ANOVA", row = "expression", response = "response"),
  base = 2, more = TRUE, addData = TRUE, ...)
}
\arguments{
\item{mat}{a matrix, i.e. gene expression}

\item{matIsX}{not implemented at this moment}

\item{meta}{meta information, a list. This can specify what type of analysis (i.e. RPPA correlated with EMT score);
row, which represents what row variables are (rows of mat); response represents what does vec mean}

\item{vec}{a vector of phenotype, categorical variable}
}
\value{
an object belonging to data frame and a class
}
\description{
Two-way ANOVA between rows of a matrix and two vectors: two main effects and then interaction effect
 
Each row is to be fitted by fn_TWanova()
}

