% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RPPAdat}
\alias{RPPAdat}
\title{An example KEGG Gene2Annotation data}
\usage{
data(RPPAdat)
}
\description{
This is an example data from TCGA Pan-cancer RPPA data (a subset)
}
\details{
Two data frames: RPPA (rows are proteins and columns are samples) and ANNO (sample information)
}
\keyword{datasets}

