% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{safelog}
\alias{safelog}
\title{log transformation of matrix, i.e. count matrix}
\usage{
safelog(dat, min = 1e-05, logfun = "log2")
}
\arguments{
\item{dat}{a vector or a matrix to be safe-log}

\item{min}{a value specifying values to be filled if <=0}

\item{logfun}{a log function to be applied, i.e. log, log10 and so on}
}
\value{
the safe-logged result, a vector or a matrix depending on the dat input
}
\description{
After specifying a minimum value, any value smaller than it would be filled with it and then applied with a log transform.
}
\details{
The user can specify a logfun for the transformation.
}

