% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_p.R
\name{saveOR}
\alias{saveOR}
\title{save() with overwrite option}
\usage{
saveOR(file, overwrite = TRUE, ..., return = FALSE)
}
\arguments{
\item{return}{default to no return}
}
\description{
save() with overwrite option
}

