% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{scatter.rowanovaMatVec}
\alias{scatter.rowanovaMatVec}
\title{Scatter plot for rowanovaMatVec class: each row produces a printted figure}
\usage{
\method{scatter}{rowanovaMatVec}(x, maxRow = 300, main = NULL,
  maintagfunc = NULL, groupShown = NULL, ...)
}
\arguments{
\item{x}{a rowanovaMatVec class (a data frame)}

\item{...}{additional parameters to scatter_rowttestMatVec}
}
\description{
Scatter plot for rowanovaMatVec class: each row produces a printted figure
}

