% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{scatter.rowcorMatVec}
\alias{scatter.rowcorMatVec}
\title{Scatter plot for rowcorMatVec class: each row produces a printted figure}
\usage{
\method{scatter}{rowcorMatVec}(x, maxRow = 1000, main = NULL,
  addLabel = FALSE, ...)
}
\arguments{
\item{x}{a rowcorMatVec class (a data frame)}

\item{addLabel}{whether to add sample names as text in the scatter plot}

\item{...}{additional parameters to scatter_rowcorMatVec}
}
\description{
Scatter plot for rowcorMatVec class: each row produces a printted figure
}

