% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{scatter.rowfisherMatVec}
\alias{scatter.rowfisherMatVec}
\title{Scatter plot for rowfisherMatVec class: each row produces a printted figure}
\usage{
\method{scatter}{rowfisherMatVec}(x, maxRow = 500, main = NULL, ...)
}
\arguments{
\item{x}{a rowfisherMatVec class (a data frame)}
}
\description{
Scatter plot for rowfisherMatVec class: each row produces a printted figure
}

