% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{scatter.rowkmMatVec}
\alias{scatter.rowkmMatVec}
\title{Scatter plot for rowkmMatVec class: each row produces a printted figure}
\usage{
\method{scatter}{rowkmMatVec}(x, maxRow = 500, main = NULL, ...)
}
\arguments{
\item{x}{a rowkmMatVec class (a data frame)}
}
\description{
Scatter plot for rowkmMatVec class: each row produces a printted figure
}

