% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{scatter.rowpairedttestMatVec}
\alias{scatter.rowpairedttestMatVec}
\title{Scatter plot for rowpairedttestMatVec class: each row produces a printted figure}
\usage{
\method{scatter}{rowpairedttestMatVec}(x, ind = NULL, maxRow = Inf, ...)

\method{scatter}{rowpairedttestMatVec}(x, ind = NULL, maxRow = Inf, ...)
}
\arguments{
\item{x}{a rowpairedttestMatVec class (a data frame)}

\item{...}{additional parameters to scatter_rowpairedttestMatVec}

\item{x}{a rowpairedttestMatVec class (a data frame)}

\item{...}{additional parameters to scatter_rowttestMatVec}
}
\description{
Scatter plot for rowpairedttestMatVec class: each row produces a printted figure

Scatter plot for rowpairedttestMatVec class: each row produces a printted figure
}

