% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{scatter.rowttestMatVec}
\alias{scatter.rowttestMatVec}
\title{Scatter plot for rowttestMatVec class: each row produces a printted figure}
\usage{
\method{scatter}{rowttestMatVec}(x, maxRow = Inf, main = NULL,
  maintagfunc = NULL, addRankP = FALSE, type = "jitter", col = NULL,
  shape = 18, ...)
}
\arguments{
\item{x}{a rowttestMatVec class (a data frame)}

\item{maintagfunc}{a function to compute additional information for main}

\item{addRankP}{whether to add rank test p value in the scatter annotation}

\item{...}{additional parameters to scatter_rowttestMatVec}
}
\description{
Scatter plot for rowttestMatVec class: each row produces a printted figure
}

