% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{scatter.rowTWanovaMatVec}
\alias{scatter.rowTWanovaMatVec}
\title{Scatter plot for rowTWanovaMatVec class: each row produces a printted figure}
\usage{
\method{scatter}{rowTWanovaMatVec}(x, var = "pval_main",
  ylabsuffix = ", expression", maxRow = 300, main = NULL, ...)
}
\arguments{
\item{x}{a rowTWanovaMatVec class (a data frame)}

\item{var}{what variable to be used to order the genes (pval_main, pval_confounder, pval_interaction, anovaP|HCC1806, ...)}

\item{ylabsuffix}{ylab suffix to be added after gene name}

\item{...}{additional parameters to plot.fn_TWanova}
}
\description{
Scatter plot for rowTWanovaMatVec class: each row produces a printted figure
}

