% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{scatter_rowanovaMatVec}
\alias{scatter_rowanovaMatVec}
\title{scatter plot for rowanovaMatVec}
\usage{
scatter_rowanovaMatVec(mat, vec, FUN, geneSel, rowanovaMatVec, matIsX = TRUE,
  maintagfunc = NULL, groupShown = NULL, ...)
}
\arguments{
\item{mat}{expression matrix used in rowanovaMatVec}

\item{vec}{vec as used in rowanovaMatVec}

\item{FUN}{plotting function}

\item{geneSel}{index for plotting; default is plotting all sorted by P}

\item{rowanovaMatVec}{rowanovaMatVec object}

\item{maintagfunc}{a function to compute additional information for main}

\item{...}{additional parameters to plotScatter}
}
\value{
multiple ggplot figures printed
}
\description{
this function needs to be polished and put into TestAndCorr_p.R
}

