% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{scatter_rowcorMatVec}
\alias{scatter_rowcorMatVec}
\title{scatter plot for rowcorMatVec}
\usage{
scatter_rowcorMatVec(mat, vec, FUN, geneSel, rowcorMatVec, smooth = FALSE,
  lm = TRUE, labels = NULL, coord_flip = FALSE, plot.margin = NULL, ...)
}
\arguments{
\item{mat}{expression matrix used in rowcorMatVec}

\item{vec}{vec as used in rowcorMatVec}

\item{FUN}{plotting function}

\item{geneSel}{index for plotting; default is plotting all sorted by P}

\item{rowcorMatVec}{rowcorMatVec object}

\item{coord_flip}{whether to flip x-y coordinate}

\item{plot.margin}{when coord_flip is true, margin may be too small; here we can adjust it}

\item{...}{additonal parameters to plotScatter}
}
\value{
multiple ggplot figures printed
}
\description{
this function needs to be polished and put into TestAndCorr_p.R
}

