% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{scatter_rowttestMatVec}
\alias{scatter_rowttestMatVec}
\title{Scatter plot for rowttestMatVec}
\usage{
scatter_rowttestMatVec(mat, vec, FUN, geneSel, rowttestMatVec,
  sizeForce = NULL, anglex = NULL, ylim = NULL, matIsX = TRUE,
  paired = FALSE, addRankP = FALSE, type = "jitter", col = NULL,
  shape = 18, maintagfunc = NULL, errorDotPlot = FALSE)
}
\arguments{
\item{mat}{expression matrix used in rowttestMatVec}

\item{vec}{vec as used in rowttestMatVec}

\item{FUN}{plotting function}

\item{geneSel}{index for plotting; default is plotting all sorted by P}

\item{rowttestMatVec}{rowttestMatVec object}

\item{type}{type to be passed to plotContCat}

\item{maintagfunc}{a function to compute additional information for main}
}
\description{
Scatter plot for rowttestMatVec
}

