% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplots_p.R
\name{selectGeneFromMeanSD}
\alias{selectGeneFromMeanSD}
\title{select genes based on mean~SD plot from plotMeanSD4mat}
\usage{
selectGeneFromMeanSD(ll, m, s)
}
\arguments{
\item{ll}{the list returned by plotMeanSD4mat()}

\item{m}{mean cutoff, usually provided in ll}

\item{s}{SD cutoff, usually provided in ll}
}
\value{
a data frame with Index, Mean and SD. Index tells what rows are selected
}
\description{
select genes based on mean~SD plot from plotMeanSD4mat
}

