% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{SIBER2}
\alias{SIBER2}
\title{Compute BI for a gene expression matrix}
\usage{
SIBER2(y, ...)
}
\arguments{
\item{y}{a vector}

\item{...}{additional parameters passed to SIBER}
}
\value{
a vector
}
\description{
Compute BI for a gene expression matrix
}

