% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myplots_p.R
\name{statContCont}
\alias{statContCont}
\title{Calculate statistics for continuous variable x and y}
\usage{
statContCont(x, y)
}
\arguments{
\item{x}{variable 1}

\item{y}{variable 2}
}
\value{
a list
}
\description{
Calculate statistics for continuous variable x and y
}

