% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{str_extract2}
\alias{str_extract2}
\title{extract string given a pattern using back reference implemented by sub}
\usage{
str_extract2(string, pattern)
}
\arguments{
\item{string}{a string vector}

\item{pattern}{a pattern that to be passed to sub}
}
\value{
returned by sub
}
\description{
this function is an improved version of str_extract where we want to match the desired substring with a pattern using back reference. 
Notice that str_extract can describe what we want, which may be difficult. Now str_extract2 describes what we don't want and what is left is the desired result.
}

