% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{subsetMainMat}
\alias{subsetMainMat}
\title{subset main expression mat}
\usage{
subsetMainMat(mat, gSel = NULL, sSel0 = NULL, cluster_rows, cluster_cols,
  sSel1 = NULL, colOrderIndex = NULL, rowOrderIndex = NULL)
}
\arguments{
\item{mat}{matrix for heatmap}

\item{gSel}{gene selection (global index or gene names). This can be index or gene names. Internally, both will be converted to integer based index (global or local)}

\item{sSel0}{sample selection filter 0; by default, this deals with augMat to remove all NA columns
the user can override this by supplying a vector of indeces or sample names.}

\item{cluster_rows}{whether to draw row clusters}

\item{cluster_cols}{whether to draw column clusters}

\item{sSel1}{sample selection (index or sample names). This is useful to select a subset of samples based
on phenotype, i.e. some mutation, like ++ to be removed}

\item{colOrderIndex}{index/sample names to order column (global index or sample names); The actual number of samples surviving is an intersect between colOrderIndex and sSel (from sSel0 and sSel1).
if not specified, column clustering will be instructed to construct; otherwise, no clumn clustering will be done later on}

\item{rowOrderIndex}{index (integer or gene names) to order rows; The actual genes surviving is an intersect between rowOrderIndex and gSel.
if not specified, row clustering will be instructed to construct; otherwise, no row clustering will be done later on}
}
\value{
expression matrix with attributes gSel, sSel and so on
}
\description{
this can be used to subset genes, extract effective columns in augMat, and further subset columns.
}
\details{
Users can subset samples through two selectors: sSel0 and sSel1. However, sSel0 is primarily preserved to select effective samples after
data augmentation. sSel1 is for subsetting. colOrderIndex can be used to sort the samples (columns). The final samples presented
is the intersection between sSel0, sSel1 and colOrderIndex. Notice that when index is provided, 
it is assumed to be global index for easy tracking (both gSel and sSel0, sSel1 and colOrderIndex). Provoding sample name or gene name
in the form of rownames or colnames also works.
}

