% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{test_tree_split}
\alias{test_tree_split}
\title{test association between tree cut and a vector}
\usage{
test_tree_split(tree, k = 3, h = NULL, vec, plot = TRUE, ...)
}
\arguments{
\item{tree}{a dendrogram}

\item{k}{k as in cutree}

\item{h}{h as in cutree}

\item{vec}{value to be tested}

\item{plot}{whether to draw a figure or just test result}

\item{...}{additional parameters passed to plotScatter}
}
\description{
test association between tree cut and a vector
}

