% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_p.R
\name{testObject}
\alias{testObject}
\title{Test if an object exist}
\usage{
testObject(object)
}
\arguments{
\item{object}{the object name, a variable}
}
\value{
TRUE or FALSE
}
\description{
Test if an object exist
}

