% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{testree.phm}
\alias{testree.phm}
\title{testree method for class phm}
\usage{
\method{testree}{phm}(obj, orientation = "column", bar = NULL, k = 3,
  h = NULL, plot = TRUE)
}
\arguments{
\item{obj}{an phm object}

\item{orientation}{specify what tree to test with, row or column}

\item{bar}{column name to specify the pheno data (column of pheno or gpheno)}

\item{k}{k as in cutree}

\item{h}{h as in cutree}

\item{plot}{whether to draw a figure or just test result}
}
\value{
test result or nothing
}
\description{
this function test row tree or column tree with specified pheno/gpheno data
}

