% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{testree}
\alias{testree}
\title{Generic testree method (for phm object produced by plotHeatmap function)}
\usage{
testree(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{other arguments}
}
\description{
Generic testree method (for phm object produced by plotHeatmap function)
}

