% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_p.R
\name{theme_base}
\alias{theme_base}
\title{a theme mimic the base system}
\usage{
theme_base(base_size = 16, base_family = "Times")
}
\arguments{
\item{base_size, }{default 16}

\item{base_family, }{default Times}
}
\description{
a theme mimic the base system
}

