% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{toFC}
\alias{toFC}
\title{given expression mean difference, transforms back to fold change: negative means down-regulate}
\usage{
toFC(vec, base = 2)
}
\arguments{
\item{vec}{a vector of difference}

\item{base}{base for transformation}
}
\value{
a vector
}
\description{
given expression mean difference, transforms back to fold change: negative means down-regulate
}

