% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{toNaturalVar}
\alias{toNaturalVar}
\title{convert R variable names (i.e. from read table with default make.names) to natural variable names}
\usage{
toNaturalVar(vars)
}
\arguments{
\item{vars}{string vector for variable names}
}
\value{
names vector of formated names
}
\description{
this replaces . with ' '; removes leading X and ending .;
}

