% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{transform_smat}
\alias{transform_smat}
\title{transform a matrix (center/scaling; truncation; clustering;) for heatmap; strongly recommends first apply subsetMainMat
to get input mat since some attributes are used}
\usage{
transform_smat(mat, scale = "row", clusterWithScaledData = FALSE,
  cluster_rows = TRUE, cluster_cols = TRUE,
  clustering_distance_rows = "correlation",
  clustering_distance_cols = "correlation", clustering_method = "ward.D",
  truncate = NA, q1 = 0, q2 = 1, Lower = NULL, Upper = NULL)
}
\arguments{
\item{mat}{a matrix, usually from subsetMainMat()}

\item{scale}{selection from c("none", "row", "column")}

\item{clusterWithScaledData}{logical indicating if use scaled data for clustering; default is FALSE}

\item{cluster_rows}{whether to cluster rows}

\item{cluster_cols}{whether to cluster columns}

\item{clustering_distance_rows}{distance metric for rows}

\item{clustering_distance_cols}{distance metric for cols}

\item{clustering_method}{clustering method}

\item{truncate}{logical indicating if truncation is needed; default is NA will enable truncate if scale!='none'}

\item{q1}{parameter q1 to truncByLimit}

\item{q2}{parameter q2 to truncByLimit}

\item{Lower}{parameter Lower to truncByQuantile()}

\item{Upper}{parameter Upper to truncByQuantile()}
}
\value{
a matrix with elements gSel, sSel, tree_row, tree_col and so on
}
\description{
transform a matrix (center/scaling; truncation; clustering;) for heatmap; strongly recommends first apply subsetMainMat
to get input mat since some attributes are used
}

