% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{truncByLimit}
\alias{truncByLimit}
\title{truncate a matrix by limit}
\usage{
truncByLimit(x, Lower, Upper)
}
\arguments{
\item{Lower}{lower cutoff value}

\item{Upper}{upper cutoff value}
}
\value{
formated matrix
}
\description{
sometimes we explicitely truncate with upper or lower limit
}

