% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{truncByQuantile}
\alias{truncByQuantile}
\title{truncate a matrix by quantile}
\usage{
truncByQuantile(x, q1, q2)
}
\arguments{
\item{q1}{lower quantile.}

\item{q2}{upper quantile.}
}
\value{
formated matrix
}
\description{
sometimes we explicitely truncate with upper or lower quantile
}

