% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_p.R
\name{truncMat}
\alias{truncMat}
\title{truncate matrix used to generate heatmap; add option to scale}
\usage{
truncMat(mat, scale = TRUE, Lower, Upper, q1 = 0.01, q2 = 0.99)
}
\arguments{
\item{mat}{the matrix to be formated}

\item{scale}{whether to scale the data before formatting}

\item{Lower}{lower cutoff value}

\item{Upper}{upper cutoff value}

\item{q1}{lower quantile. It is effective only if Lower is not specified}

\item{q2}{upper quantile. It is effective only if Lower is not specified}
}
\value{
formated matrix
}
\description{
this is especially useful for heatmap
}

