% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{updatePal}
\alias{updatePal}
\title{update color pal (as input for prepcolbar)}
\usage{
updatePal(oldpal, newpal = NULL)
}
\arguments{
\item{oldpal}{a list of colors, usually returned by autocolpal4DF}

\item{newpal}{a list of colors, usually user specified; default is NULL, which means no updates at all.}
}
\value{
a named list of colors
}
\description{
given an old pal and a list of newly specified pal, update when compatible; when no shared names, no updates will be made
}

