% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aheatmat.R
\name{vec2colvec}
\alias{vec2colvec}
\title{convert a character vector to a color vector by specified col.}
\usage{
vec2colvec(vec, colpal = NULL)
}
\arguments{
\item{vec}{a vector of values. it can be numeric or categorical.}

\item{colpal}{a color vector. If NA is present, it needs to to be taken into account.}
}
\value{
a color vecotor same length as vec
}
\description{
convert a character vector to a color vector by specified col.
}

