% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{venndiagram}
\alias{venndiagram}
\title{Generic venndiagram method}
\usage{
venndiagram(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{other arguments}
}
\description{
Generic venndiagram method
}

