% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{venndiagram.rowTWanovaMatVec}
\alias{venndiagram.rowTWanovaMatVec}
\title{venndiagram method for rowTWanovaMatVec class}
\usage{
\method{venndiagram}{rowTWanovaMatVec}(x, FDR = NULL, selVar = "gSelectInd",
  pcutoff = 0.05, oma = c(6, 2, 1, 2), use = NULL, cat.pos = 0,
  cat.dist = 0.03, main = "", ...)
}
\arguments{
\item{x}{a rowTWanovaMatVec class (a data frame)}

\item{use}{either 'venn2' or 'imageVenn'; default is NULL, in which case 'venn2' would be selected if less or equal to 5 cell lines}

\item{cat.pos}{Vector giving the position (in degrees) of each category name along the circle, with 0 at 12 o'clock, being on top}

\item{cat.dist}{Vector giving the distance (in npc units) of each category name from the edge of the circle (can be negative)}

\item{...}{additional parameters to plot function}
}
\description{
When number of cell lines is smaller than 5, venn2 would be used; otherwise, imageVenn would be used to
display genes selected at a given method (as selected using create_bum)
}

