% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAndCorr_p.R
\name{wilcoxP}
\alias{wilcoxP}
\title{wilcox P value for a vector of response and a vector of group}
\usage{
wilcoxP(data, class, switchXY = FALSE, alternative = "two.sided", ...)
}
\arguments{
\item{data}{a vector of y}

\item{class}{a vector of group (binary categorical)}

\item{switchXY}{sometimes we have data as binary and class as continuous (for some reason, we do not want to change this in calling ttestP) and we need to swtich}

\item{alternative}{types of test}

\item{...}{additional parameters passed to wilcox.test}
}
\value{
the P value
}
\description{
If both x and y are given and paired is FALSE, a Wilcoxon rank sum test (equivalent to the Mann-Whitney test: see the Note) is carried out.
}

