% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_p.R
\name{WriteXLS_arrayOR}
\alias{WriteXLS_arrayOR}
\title{write xls (using WriteXLS) with overwrite option from an array
 default slice in margin 1, which means different rows are stored into different sheets}
\usage{
WriteXLS_arrayOR(file, overwrite = TRUE, transpose = TRUE, ...,
  return = FALSE)
}
\description{
write xls (using WriteXLS) with overwrite option from an array
 default slice in margin 1, which means different rows are stored into different sheets
}

