% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_p.R
\name{WriteXLS_listOR}
\alias{WriteXLS_listOR}
\title{write xls (using WriteXLS) with overwrite option from a list}
\usage{
WriteXLS_listOR(file, overwrite = TRUE, ..., return = FALSE)
}
\arguments{
\item{return}{default to no return}
}
\description{
write xls (using WriteXLS) with overwrite option from a list
}

