% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_p.R
\name{WriteXLSOR}
\alias{WriteXLSOR}
\title{write xls (using WriteXLS) with overwrite option}
\usage{
WriteXLSOR(file, overwrite = TRUE, ..., return = FALSE)
}
\arguments{
\item{return}{default to no return}
}
\description{
write xls (using WriteXLS) with overwrite option
}

